Add feature mark a book as currently being read
Create methods of the Library class to implement new feature.
The currently_readin_book() method takes a string as an argumnet and
pass it to the _get_book_by_title() helper method that returns a Book
instance representing a book previously added to the collection.

Call the currently_reading_book() Library method from
"reading_progress.py" when appropriate CLI argument is given.