Create setter method for current_page attribute
The method bookmark_page() attempts to update the current_page attribute
with the page number given as argument.
If the chapter of the page attempted to be bookmarked is already
finished the _get_first_unfinished_chapter_name() helper method is
called and the first page of the chapter returned is bookmarked.