Fix method changes not being saved
The class's method to write to the json file hade its name and behaviour
changed after the add_book() and set_currently_reading_book() methods
were created.
Updated the add_book() and set_currently_reading_book() methods to work
with the archive book() method.

to work with the archive_book()