Fix not being able to finish last chapter
If the bookmarked page is in the chapter being marked as finished the
bookmark automaticaly is moved to the first page of the first chapter
identified as unfinished.
The bug happens when an attempt to mark the last unfinished chapter as
finished is being made since there is nowhere to move the bookmark to.

To circumvent this the current_page attribute is set to None in case
there are nu unfinished chapters to place the bookmark in.
Various flow control and error handling statements have been implemented
to enable the program to run and display correct information when the
current_page attribute is set to none.