Fix completeion message when on first page
The if else statement to print a chapter progress or book completeion
message was checking for a percentage of chapter completion causing the
book completeion message to be printed if the first page of the book was
bookmarked.
Changed the if statement to check for boolean value of the
currently_read_chapter variable.